requirements.txt
google-genai
python-dotenv
pynput
pyperclip
customtkinter
packaging
